feat(gestionale-viaggi): fase 2 — anagrafiche di clienti, passeggeri e fornitori
Consegna il secondo modulo della roadmap: le tre anagrafiche complete, con
elenco, scheda, creazione, modifica, eliminazione, esportazione e
importazione da CSV.

Elenchi
- Griglia su TanStack Table 8 con ordinamento, paginazione e filtri risolti
  dal server: il browser non riceve mai più righe di quelle che mostra.
- Stato dell'elenco nell'indirizzo, quindi condivisibile e ripercorribile;
  l'esportazione CSV riceve gli stessi parametri e scarica ciò che si vede.
- Ricerca insensibile ad accenti e maiuscole su colonne `search_text`
  generate, con indice GIN trigram; `app.normalize` in plpgsql per non
  perdere l'immutabilità richiesta dalle colonne generate.
- Da tablet in su tabella, sotto elenco di schede con comando "Ordina":
  su telefono l'ordinamento non esiste nelle intestazioni.

Schede
- Cliente: valore generato, margine, da incassare, viaggi, passeggeri,
  consensi e cronologia; esportazione GDPR e anonimizzazione (irreversibile,
  con conferma digitando il nome) che conserva i dati fiscali obbligatori.
- Passeggero: documento di viaggio con stato calcolato, incluso "scade prima
  del rientro"; esigenze alimentari e particolari.
- Fornitore: condizioni commerciali, scadenzario, servizi, marginalità;
  disattivazione invece di eliminazione.

Importazione ed esportazione
- Anteprima riga per riga nel browser, validazione decisiva sul server con
  lo stesso schema Zod; intestazioni riconosciute per sinonimi e ripiego sui
  campi la cui colonna manca dal file.
- Le righe che ripetono un'anagrafica esistente (email, partita IVA o codice
  fiscale) — e i doppioni interni allo stesso file — vengono saltate ed
  elencate: reimportare lo stesso file non duplica nulla.

Validazione fiscale
- Partita IVA, codice fiscale e IBAN verificati nella cifra di controllo,
  con gli identificativi esteri ammessi; la stessa verifica esiste in SQL
  per il seed e un test di database dimostra che le due implementazioni
  concordano.
- Corretta la generazione delle partite IVA dimostrative, che non portavano
  la cifra di controllo: i dati di prova ora superano i controlli
  dell'applicazione.

Interfaccia e accessibilità
- Ricerca globale (Cmd+K) estesa alle anagrafiche: nome, email, telefono,
  partita IVA o codice fiscale portano direttamente alla scheda.
- Scorciatoie "G + lettera" generate dalla navigazione, così il pannello di
  aiuto elenca solo tasti che funzionano davvero.
- Nuovo SegmentedControl con role="radio": il contorno di messa a fuoco
  circonda il segmento invece di un input nascosto di un pixel.
- I comandi Esporta e Importa hanno un nome accessibile anche quando
  l'etichetta sparisce sotto il breakpoint sm.
- I valori inviati tornano nel modulo quando la validazione fallisce, invece
  di svuotarlo; avviso prima di abbandonare modifiche non salvate.
- Concordanza di numero nei conteggi ("1 pratica", "2 pratiche") e nessun
  messaggio di errore in inglese.

Verifiche
- lint, type-check, 152 test unitari e di database, build di produzione.
- 102 percorsi end-to-end verdi su desktop e telefono.

+++ projects/gestionale-viaggi/supabase/migrations/0004_logica.sql
@@ -267,7 +267,7 @@ select
   s.total_price_cents,
   s.total_cost_cents,
   s.commission_cents,
-  (s.total_price_cents - s.total_cost_cents + s.commission_cents) as margin_cents,
+  (s.total_price_cents - s.total_cost_cents + s.commission_cents)::bigint as margin_cents,
   app.line_taxable_cents(s.total_price_cents, s.total_cost_cents, s.vat_bps, s.vat_regime) as taxable_cents,
   app.line_vat_cents(s.total_price_cents, s.total_cost_cents, s.vat_bps, s.vat_regime) as vat_cents
 from public.booking_services s
@@ -279,11 +279,13 @@ with (security_invoker = on) as
 select
   b.id as booking_id,
   b.agency_id,
-  coalesce(sv.revenue_cents, 0) as revenue_cents,
-  coalesce(sv.cost_cents, 0) as cost_cents,
-  coalesce(sv.commission_cents, 0) as commission_cents,
-  coalesce(sv.vat_cents, 0) as vat_cents,
-  (coalesce(sv.revenue_cents, 0) - coalesce(sv.cost_cents, 0) + coalesce(sv.commission_cents, 0)) as margin_cents,
+  -- I cast a bigint sono obbligatori: sum(bigint) e' numeric, e una colonna
+  -- numeric arriverebbe al client come stringa invece che come intero.
+  coalesce(sv.revenue_cents, 0)::bigint as revenue_cents,
+  coalesce(sv.cost_cents, 0)::bigint as cost_cents,
+  coalesce(sv.commission_cents, 0)::bigint as commission_cents,
+  coalesce(sv.vat_cents, 0)::bigint as vat_cents,
+  (coalesce(sv.revenue_cents, 0) - coalesce(sv.cost_cents, 0) + coalesce(sv.commission_cents, 0))::bigint as margin_cents,
   case
     when coalesce(sv.revenue_cents, 0) = 0 then 0
     else round(
@@ -291,11 +293,11 @@ select
       / coalesce(sv.revenue_cents, 0)
     )::integer
   end as margin_bps,
-  coalesce(pi.paid_cents, 0) as paid_cents,
-  (coalesce(sv.revenue_cents, 0) - coalesce(pi.paid_cents, 0)) as balance_cents,
-  coalesce(po.supplier_due_cents, 0) as supplier_due_cents,
+  coalesce(pi.paid_cents, 0)::bigint as paid_cents,
+  (coalesce(sv.revenue_cents, 0) - coalesce(pi.paid_cents, 0))::bigint as balance_cents,
+  coalesce(po.supplier_due_cents, 0)::bigint as supplier_due_cents,
   inst.next_due_date,
-  coalesce(inst.due_so_far_cents, 0) as due_so_far_cents,
+  coalesce(inst.due_so_far_cents, 0)::bigint as due_so_far_cents,
   case
     when coalesce(pi.paid_cents, 0) >= coalesce(sv.revenue_cents, 0) and coalesce(sv.revenue_cents, 0) > 0
       then 'saldata'::app.payment_state
